amend! trace_fprobe: update for cheri
trace_fprobe: update for cheri

Update trace_fprobe for cheri. This is the code that creates fprobes from
input strings written to /sys/kernel/tracing/dynamic_events.

Treat the stack as an array of capabilities in
ftrace_regs_get_kernel_stack_nth.

Fix a size check for the shadow stack. On cheri, it is a uintptr_t array.

The fentry_trace_entry_head and fexit_trace_entry_head structs are stored
in the ringbuffer. Addresses in those structs must be downgraded.

The parsing of dynamic_events' input strings creates a sequence of fetch
operations which are processed in four stages.

We use a capability in stage 1 where data is read from the stack, from a
register or from struct task_struct. The metadata aren't used at the
moment, we're planning to pass them to the generated dynamic events in
the future.

edata in stage 1 is an array of some function arguments that were saved
at function entry to be processed at exit. edata uses the shadow stack,
which is a uintptr_t array on cheri. We have to save the arguments as
capabilities to support different fetch types when a saved argument is
processed later (something like 'f my_fct%return $arg1:cap').

Stages 2-4 may access memory at a user-provided address + offset. Our only
option is to fabricate a capability for each access.

+++ kernel/trace/trace_probe.c
@@ -904,7 +904,7 @@ static int __store_entry_arg(struct trace_probe *tp, int argnum)
 		last_offset = earg->code[i - 1].offset;
 	}
 
-	offset = last_offset + sizeof(unsigned long);
+	offset = last_offset + sizeof(uintptr_t);
 	store_entry_arg_at(&earg->code[i], argnum, offset);
 	return offset;
 }
@@ -916,7 +916,7 @@ int traceprobe_get_entry_data_size(struct trace_probe *tp)
 	if (!earg)
 		return 0;
 
-	return get_entry_arg_max_offset(earg) + sizeof(unsigned long);
+	return get_entry_arg_max_offset(earg) + sizeof(uintptr_t);
 }
 
 void store_trace_entry_data(void *edata, struct trace_probe *tp, struct pt_regs *regs)
